# specs/001-ga-ticketing/contracts/api.yaml
openapi: 3.0.3
info:
  title: GA Ticketing System API
  description: General Affairs ticketing system for service requests, approvals, and asset management
  version: 1.0.0
  contact:
    name: GA Team
    email: ga@company.com
servers:
  - url: https://api.company.com/v1
    description: Production server
  - url: https://staging-api.company.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
paths:
...
  /tickets:
    get:
      summary: Get list of tickets based on user role
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/CategoryFilter'
        - $ref: '#/components/parameters/PriorityFilter'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Tickets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create a new service request ticket
      tags:
        - Tickets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tickets/{ticketId}:
    get:
      summary: Get ticket details by ID
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Ticket retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      summary: Update ticket details
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdateRequest'
      responses:
        '200':
          description: Ticket updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /assets:
    get:
      summary: Get list of assets (Admin only)
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetCategoryFilter'
        - $ref: '#/components/parameters/AssetConditionFilter'
        - $ref: '#/components/parameters/LocationFilter'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Assets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create new asset (Admin only)
      tags:
        - Assets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreateRequest'
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /assets/{assetId}:
    get:
      summary: Get asset details (Admin only)
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Asset retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      summary: Update asset details (Admin only)
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetUpdateRequest'
      responses:
        '200':
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the ticket
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the asset
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, waiting_approval, approved, rejected, in_progress, completed, closed]
      description: Filter tickets by status
    CategoryFilter:
      name: category
      in: query
      schema:
        type: string
        enum: [office_supplies, facility_maintenance, pantry_supplies, meeting_room, office_furniture, general_service]
      description: Filter tickets by category
    PriorityFilter:
      name: priority
      in: query
      schema:
        type: string
        enum: [low, medium, high]
      description: Filter tickets by priority
    AssetCategoryFilter:
      name: category
      in: query
      schema:
        type: string
        enum: [office_furniture, office_supplies, pantry_supplies, facility_equipment, meeting_room_equipment, cleaning_supplies]
      description: Filter assets by category
    AssetConditionFilter:
      name: condition
      in: query
      schema:
        type: string
        enum: [good, needs_maintenance, broken]
      description: Filter assets by condition
    LocationFilter:
      name: location
      in: query
      schema:
        type: string
      description: Filter assets by location
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [requester, approver, admin]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # Ticket schemas
    TicketCreateRequest:
      type: object
      required:
        - title
        - description
        - category
        - priority
        - estimated_cost
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        category:
          $ref: '#/components/schemas/TicketCategory'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        estimated_cost:
          type: integer
          minimum: 0
          description: Estimated cost in Indonesian Rupiah
    TicketUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        actual_cost:
          type: integer
          minimum: 0
          description: Actual cost in Indonesian Rupiah
...
    TicketResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_number:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/TicketCategory'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        requester:
          $ref: '#/components/schemas/UserResponse'
        assigned_admin:
          $ref: '#/components/schemas/UserResponse'
        estimated_cost:
          type: integer
        actual_cost:
          type: integer
        requires_approval:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
    TicketListResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Comment schemas
...
    CommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
...
    AssetCreateRequest:
      type: object
      required:
        - name
        - description
        - category
        - quantity
        - location
        - unit_cost
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        category:
          $ref: '#/components/schemas/AssetCategory'
        quantity:
          type: integer
          minimum: 1
        location:
          type: string
          maxLength: 255
        unit_cost:
          type: integer
          minimum: 0
          description: Unit cost in Indonesian Rupiah
    AssetUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        location:
          type: string
          maxLength: 255
        condition:
          $ref: '#/components/schemas/AssetCondition'
        unit_cost:
          type: integer
          minimum: 0
        last_maintenance_at:
          type: string
          format: date-time
        next_maintenance_at:
          type: string
          format: date-time
...
    AssetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_code:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/AssetCategory'
        quantity:
          type: integer
        available_quantity:
          type: integer
        location:
          type: string
        condition:
          $ref: '#/components/schemas/AssetCondition'
        unit_cost:
          type: integer
        last_maintenance_at:
          type: string
          format: date-time
        next_maintenance_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AssetListResponse:
      type: object
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/AssetResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Enum schemas
    TicketCategory:
      type: string
      enum: [office_supplies, facility_maintenance, pantry_supplies, meeting_room, office_furniture, general_service]
    TicketPriority:
      type: string
      enum: [low, medium, high]
    TicketStatus:
      type: string
      enum: [pending, waiting_approval, approved, rejected, in_progress, completed, closed]
    AssetCategory:
      type: string
      enum: [office_furniture, office_supplies, pantry_supplies, facility_equipment, meeting_room_equipment, cleaning_supplies]
    AssetCondition:
      type: string
      enum: [good, needs_maintenance, broken]
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...